requirements.txt
pymilvus
psycopg2-binary
pypdf2
pytesseract
Pillow
pdf2image
sentence-transformers
transformers
torch
open-clip-torch
langchain
langchain-community
rank-bm25
llama-cpp-python
gradio
numpy
tqdm
opencv-python-headless
